rename query.py to program_query.py
